Added modularized database connection
I've included a basic database connection with Mongoose. It is
outsourced in the model folder, which holds the database configuration
(db.js) and any future additional custom mongoose schemas.

Currently the databas-uri is hardcoded, but the code is flexible enough
so that the uri can easily be changed when needed, or sourced from a
potential configuration file.

This part is included into the main app.js as db.